services/notifier/pyproject.toml
[project]
name = "notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "python-telegram-bot>=22.2",
    "uvicorn>=0.35.0",
]
